Cleaned controllers, used transaction managment on services, optimizated entities loading, used BigDecimal

+++ src/main/java/com/andlvovsky/periodicals/service/OrderService.java
@@ -2,9 +2,11 @@ package com.andlvovsky.periodicals.service;
 
 import com.andlvovsky.periodicals.model.basket.Basket;
 
+import java.math.BigDecimal;
+
 public interface OrderService {
 
-    double calculateCost(Basket basket);
+    BigDecimal calculateCost(Basket basket);
 
     void registerOrder(Basket basket);
